pyproject.toml
[tool.poetry]
name = "temba"
version = "7.4.2"
description = "Hosted service for visually building interactive messaging applications"
authors = ["Nyaruka <code@nyaruka.com>"]
license = "AGPL-3"

[tool.poetry.urls]
repository = "http://github.com/rapidpro/rapidpro"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.6"
django-compressor = "^3.1"
django-countries = "^7.0"
django-flatpickr = "^1.0.0"
django-hamlpy = "^1.4.4"
django-mptt = "^0.12.0"
django-redis = "^4.12.1"
django-storages = "^1.11.1"
django-timezone-field = "^4.1.1"
djangorestframework = "^3.12.2"
dj-database-url = "^0.5.0"
smartmin = "^4.1.0"
celery = "^5.1.0"
redis = "^3.5.3"
elasticsearch = "^7.11.0"
elasticsearch-dsl = "^7.3.0"
analytics-python = "^1.2.9"
boto3 = "^1.17.21"
cryptography = "3.4.8"
vonage = "2.5.2"
pyfcm = "^1.5.1"
pyotp = "2.4.1"
python-telegram-bot = "^11.1.0"
sentry-sdk = "^1.0.0"
stripe = "^2.56.0"
twilio = "6.24.0"
twython = "3.5.0"
geojson = "^2.5.0"
Markdown = "^3.3.4"
polib = "^1.1.0"
pyexcel = "^0.6.6"
pyexcel-xls = "^0.6.2"
pyexcel-xlsx = "^0.6.0"
python-magic = "^0.4.22"
xlsxlite = "^0.2.0"
colorama = "^0.4.4"
gunicorn = "^20.0.4"
iptools = "^0.7.0"
iso-639 = "^0.4.5"
iso8601 = "^0.1.14"
phonenumbers = "*"
pycountry = "^20.7.3"
python-dateutil = "2.2"
cffi = "1.15.0"
psycopg2-binary = "^2.9.1"
python-intercom = "^3.1.0"
packaging = "^20.9"
requests-toolbelt = "^0.9.1"
chardet = "^4.0.0"
openpyxl = "^3.0.9"
python-json-logger = "2.0.0"
#----- engage
bandwidth-sdk = "3.1.0"
iris_sdk = {git = "https://github.com/bandwidthcom/python-bandwidth-iris"}
amqp = "2.6.1"
dicttoxml = "1.7.4"
django-getenv = "1.3.1"
django-cache-url = "3.4.2"
pypng = "^0.20220715.0"
pyqrcode = "^1.2"
uwsgi = "2.0.18"
whitenoise = "^5.1"
#----- oidc _PROVIDER_
#pyjwt = "^2.3.0" #required by oauthlib
#oauthlib = "^3.2.0" #required by django-oauth-toolkit
#django-oauth-toolkit = "^1.7.1"
#django-cors-headers = "^3.11.0"
#----- oidc _CLIENT_
django-oauth2-authcodeflow = "0.5.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
coverage = "^5.5"
flake8 = "^4.0.1"
isort = "^5.10.1"
codecov = "^2.1.11"
responses = "^0.12.1"
beautifulsoup4 = "^4.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
